requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib>=3.8.0
numpy>=1.26.0
psycopg2-binary==2.9.10
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.35
thyagomotalib3==0.0.1
typing_extensions==4.12.2
urllib3==2.2.0
Werkzeug==3.0.4
WTForms==3.1.2
zipp==3.21.0